# Refactor SDK span creation and implementation