--- a/sdk/trace/tracer.go
+++ b/sdk/trace/tracer.go
@@ -17,6 +17,7 @@ package trace // import "go.opentelemetry.io/otel/sdk/trace"
 import (
 	"context"
 	rt "runtime/trace"
+	"time"
 
 	"go.opentelemetry.io/otel/trace"
 
@@ -34,8 +35,7 @@ var _ trace.Tracer = &tracer{}
 //
 // The Span is created with the provided name and as a child of any existing
 // span context found in the passed context. The created Span will be
-// configured appropriately by any SpanOption passed. Any Timestamp option
-// passed will be used as the start time of the Span's life-cycle.
+// configured appropriately by any SpanOption passed.
 func (tr *tracer) Start(ctx context.Context, name string, options ...trace.SpanStartOption) (context.Context, trace.Span) {
 	config := trace.NewSpanStartConfig(options...)
 
@@ -46,30 +46,108 @@ func (tr *tracer) Start(ctx context.Context, name string, options ...trace.SpanS
 		}
 	}
 
-	span := startSpanInternal(ctx, tr, name, config)
-	for _, l := range config.Links() {
-		span.addLink(l)
-	}
-	span.SetAttributes(config.Attributes()...)
-
-	span.tracer = tr
-
-	if span.IsRecording() {
+	s := tr.newSpan(ctx, name, config)
+	if s.IsRecording() {
 		sps, _ := tr.provider.spanProcessors.Load().(spanProcessorStates)
 		for _, sp := range sps {
-			sp.sp.OnStart(ctx, span)
+			sp.sp.OnStart(ctx, s)
 		}
 	}
 
-	ctx, span.executionTracerTaskEnd = func(ctx context.Context) (context.Context, func()) {
-		if !rt.IsEnabled() {
-			// Avoid additional overhead if
-			// runtime/trace is not enabled.
-			return ctx, func() {}
-		}
-		nctx, task := rt.NewTask(ctx, name)
-		return nctx, task.End
-	}(ctx)
+	ctx, s.executionTracerTaskEnd = newRuntimeTask(ctx, name)
 
-	return trace.ContextWithSpan(ctx, span), span
+	return trace.ContextWithSpan(ctx, s), s
+}
+
+// newSpan returns a new configured span.
+func (tr *tracer) newSpan(ctx context.Context, name string, config *trace.SpanConfig) *span {
+	// If told explicitly to make this a new root use a zero value SpanContext
+	// as a parent which contains an invalid trace ID and is not remote.
+	var psc trace.SpanContext
+	if config.NewRoot() {
+		ctx = trace.ContextWithSpanContext(ctx, psc)
+	} else {
+		psc = trace.SpanContextFromContext(ctx)
+	}
+
+	// If there is a valid parent trace ID, use it to ensure the continuity of
+	// the trace. Always generate a new span ID so other components can rely
+	// on a unique span ID, even if the Span is non-recording.
+	var tid trace.TraceID
+	var sid trace.SpanID
+	if !psc.TraceID().IsValid() {
+		tid, sid = tr.provider.idGenerator.NewIDs(ctx)
+	} else {
+		tid = psc.TraceID()
+		sid = tr.provider.idGenerator.NewSpanID(ctx, tid)
+	}
+
+	s := new(span)
+	s.attributes = newAttributesMap(tr.provider.spanLimits.AttributeCountLimit)
+	s.events = newEvictedQueue(tr.provider.spanLimits.EventCountLimit)
+	s.links = newEvictedQueue(tr.provider.spanLimits.LinkCountLimit)
+	s.spanLimits = tr.provider.spanLimits
+
+	samplingResult := tr.provider.sampler.ShouldSample(SamplingParameters{
+		ParentContext: ctx,
+		TraceID:       tid,
+		Name:          name,
+		Kind:          config.SpanKind(),
+		Attributes:    config.Attributes(),
+		Links:         config.Links(),
+	})
+
+	scc := trace.SpanContextConfig{
+		TraceID:    tid,
+		SpanID:     sid,
+		TraceState: samplingResult.Tracestate,
+	}
+	if isSampled(samplingResult) {
+		scc.TraceFlags = psc.TraceFlags() | trace.FlagsSampled
+	} else {
+		scc.TraceFlags = psc.TraceFlags() &^ trace.FlagsSampled
+	}
+	s.spanContext = trace.NewSpanContext(scc)
+
+	if !isRecording(samplingResult) {
+		return s
+	}
+
+	startTime := config.Timestamp()
+	if startTime.IsZero() {
+		startTime = time.Now()
+	}
+	s.startTime = startTime
+
+	s.spanKind = trace.ValidateSpanKind(config.SpanKind())
+	s.name = name
+	s.parent = psc
+	s.resource = tr.provider.resource
+	s.instrumentationLibrary = tr.instrumentationLibrary
+	s.tracer = tr
+
+	s.SetAttributes(samplingResult.Attributes...)
+	s.SetAttributes(config.Attributes()...)
+
+	for _, l := range config.Links() {
+		s.addLink(l)
+	}
+
+	return s
+}
+
+// newRuntimeTask starts a runtime.Task with the passed name and returns both
+// a context containing the task and an end function for the task.
+//
+// If the runtime tracing is not enabled the original context is returned with
+// an empty function to call for end.
+func newRuntimeTask(ctx context.Context, name string) (context.Context, func()) {
+	if !rt.IsEnabled() {
+		// Avoid additional overhead if
+		// runtime/trace is not enabled.
+		return ctx, func() {}
+	}
+	nctx, task := rt.NewTask(ctx, name)
+	return nctx, task.End
+
 }