The span creation and configuration process is split across the tracer Start method and the startSpanInternal function, each living in different files. This adds confusion when developing. It requires the developer to remember certain parts of the configuration happen in one place or the other. Additionally, when it is determined that a span is not recorded during its creation it still is implemented by a full functioning `Span` implementation. This adds unneeded memory overhead. These changes unify the creation and configuration of a new span and add a new `nonRecordingSpan` type.

Unification of the span creation in a newSpan method on the tracer inherently associates the tracer with the span creation and centralizes all the configuration needed for a new span. This creation path is split in the newSpan method. One for a non-recording and another for a recording span.

# Benchmarks

Benchmarks performed in the `otel/sdk/trace` package.

## Before

```sh
$ go test -bench=.
go.opentelemetry.io/otel/sdk/trace.recordStackTrace(...)
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/sdk/trace
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
BenchmarkAttributesMapToKeyValue-8     	  368532	      3135 ns/op
BenchmarkStartEndSpan/AlwaysSample-8   	 1372064	       872.5 ns/op	     848 B/op	       9 allocs/op
BenchmarkStartEndSpan/NeverSample-8    	 1588966	       741.9 ns/op	     752 B/op	       8 allocs/op
BenchmarkSpanWithAttributes_4/AlwaysSample-8         	  714387	      1647 ns/op	    1584 B/op	      17 allocs/op
BenchmarkSpanWithAttributes_4/NeverSample-8          	 1367698	       877.7 ns/op	     944 B/op	       9 allocs/op
BenchmarkSpanWithAttributes_8/AlwaysSample-8         	  551848	      2054 ns/op	    2112 B/op	      23 allocs/op
BenchmarkSpanWithAttributes_8/NeverSample-8          	 1257816	       953.5 ns/op	    1136 B/op	       9 allocs/op
BenchmarkSpanWithAttributes_all/AlwaysSample-8       	  557142	      1965 ns/op	    1936 B/op	      21 allocs/op
BenchmarkSpanWithAttributes_all/NeverSample-8        	 1216846	       931.6 ns/op	    1072 B/op	       9 allocs/op
BenchmarkSpanWithAttributes_all_2x/AlwaysSample-8    	  323812	      3457 ns/op	    3236 B/op	      32 allocs/op
BenchmarkSpanWithAttributes_all_2x/NeverSample-8     	 1000000	      1177 ns/op	    1392 B/op	       9 allocs/op
BenchmarkTraceID_DotString-8                         	14290009	        80.33 ns/op
BenchmarkSpanID_DotString-8                          	19520773	        59.88 ns/op
PASS
```

## After

```sh
$ go test -bench=.
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/sdk/trace
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
BenchmarkAttributesMapToKeyValue-8     	  382208	      2932 ns/op
BenchmarkStartEndSpan/AlwaysSample-8   	 1284544	       911.1 ns/op	     848 B/op	       9 allocs/op
BenchmarkStartEndSpan/NeverSample-8    	 3394278	       341.1 ns/op	     224 B/op	       3 allocs/op
BenchmarkSpanWithAttributes_4/AlwaysSample-8         	  690414	      1576 ns/op	    1584 B/op	      17 allocs/op
BenchmarkSpanWithAttributes_4/NeverSample-8          	 2494444	       474.5 ns/op	     416 B/op	       4 allocs/op
BenchmarkSpanWithAttributes_8/AlwaysSample-8         	  495074	      2243 ns/op	    2112 B/op	      23 allocs/op
BenchmarkSpanWithAttributes_8/NeverSample-8          	 2070771	       569.0 ns/op	     608 B/op	       4 allocs/op
BenchmarkSpanWithAttributes_all/AlwaysSample-8       	  550532	      2135 ns/op	    1936 B/op	      21 allocs/op
BenchmarkSpanWithAttributes_all/NeverSample-8        	 2279611	       539.2 ns/op	     544 B/op	       4 allocs/op
BenchmarkSpanWithAttributes_all_2x/AlwaysSample-8    	  334038	      3591 ns/op	    3236 B/op	      32 allocs/op
BenchmarkSpanWithAttributes_all_2x/NeverSample-8     	 1617702	       742.9 ns/op	     864 B/op	       4 allocs/op
BenchmarkTraceID_DotString-8                         	13296470	        80.64 ns/op
BenchmarkSpanID_DotString-8                          	19072332	        62.09 ns/op
PASS
```