@jmacd 2021-08-31
(This is a nice optimization!)